Add hierarchical PrefixCache with host DRAM
Support hierarchical cache including host DRAM in PrefixCache.
It is useful when the transfer latency is lower than the recompute common prefix KV cache.
Assume caches in the HBM are the subset in the DRAM.
The cache size need to fit in the HBM.

Add specific devices used in HBMStorage.
Add retrieve to specific devices.

Possible future improvements:
* Benchmark and add config to decide if loading from DRAM is beneficial depend on common prefix length.
* Add a config to disable DRAM cache if HBM is larger enough to get the beneficial without use DRAM.

b/397854783