Init encoder_images only when use_multimodal=True
Current code initializes a numpy array on the local JAX client which scales as the number of slices increases (because of dependency on the global batch size). This happens even for non image use cases.

Fixing this such that it is more inline with the PR that introduced this change: https://github.com/AI-Hypercomputer/maxtext/commit/2fd27e27f141199f2529a90fa03ae0aae53332e4