[maxtext] improve profiling in decode
Limit the profiling of decode to `config.profiler_steps` steps. User may request generate many tokens, but since each generation is identical, it's not necessary to profile all of them. Profiling only the first few should be sufficient.

Also added a method in Profiler to do post-pocessing on the collected trace.

PiperOrigin-RevId: 775799587