Switch from pytest-split to pytest-xdist for parallel test execution.
Previously, CPU tests are distributed across multiple workers using pytest-split, which assigns the same number of tests to each worker. However, since the runtime of tests is different, some workers end up finishing fast and stand idle while others take a long time, so we're not utilizing the workers fully.
This change replaces the use of pytest-split with pytest-xdist which dynamically assigns work to workers.